+++ src/context/toast.test.tsx
@@ -1,7 +1,7 @@
 import { renderHook } from "@testing-library/react-hooks";
 import { render, fireEvent, waitFor, act } from "test_utils/test-utils";
 import { RequireAtMostOne } from "types/utils";
-import { RenderFakeToastContext } from "./__tests__/toast";
+import { RenderFakeToastContext } from "./__mocks__/toast";
 import { useToastContext, ToastProvider } from "./toast";
 
 type useToastContextProps = ReturnType<typeof useToastContext>;
